backend/requirements.txt
# Backend dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy[asyncio]==2.0.32
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
pgvector==0.3.2

# AI / ML
langchain==0.2.16
langchain-community==0.2.16
langchain-ollama==0.1.3
sentence-transformers==3.0.1
transformers==4.44.0
torch==2.4.0
Pillow==10.4.0

# NLP
nltk==3.9.1
spacy==3.7.6

# OCR
# qwen-vl-utils==0.0.2  # Uncomment when using QWEN VL

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
uuid6==2024.7.10

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.0
